deepsolv_assignment/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
pydantic==2.5.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
lxml==4.9.3
httpx==0.25.2